+++ src/hike/messages/__init__.py
@@ -3,6 +3,7 @@
 ##############################################################################
 # Local imports.
 from .clipboard import CopyToClipboard
+from .command_line import HandleInput
 from .history import ClearHistory, RemoveHistoryEntry
 from .local_view import SetLocalViewRoot
 from .opening import OpenFrom, OpenFromForge, OpenFromHistory, OpenLocation
@@ -12,6 +13,7 @@ from .opening import OpenFrom, OpenFromForge, OpenFromHistory, OpenLocation
 __all__ = [
     "ClearHistory",
     "CopyToClipboard",
+    "HandleInput",
     "OpenFrom",
     "OpenFromForge",
     "OpenFromHistory",