+++ src/hike/widgets/command_line/general.py
@@ -1,6 +1,11 @@
 """Provides general application commands for the command line."""
 
 ##############################################################################
+# Python imports.
+from functools import partial
+from typing import Callable
+
+##############################################################################
 # Textual imports.
 from textual.message import Message
 from textual.widget import Widget
@@ -17,6 +22,7 @@ from ...commands import (
     JumpToLocalBrowser,
     JumpToTableOfContents,
 )
+from ...messages import HandleInput
 from .base_command import InputCommand
 
 
@@ -96,6 +102,15 @@ class QuitCommand(GeneralCommand):
     COMMAND = "`quit`"
     ALIASES = "`q`"
     MESSAGE = Quit
+
+
+##############################################################################
+class ChangeLogCommand(GeneralCommand):
+    """Show Hike's ChangeLog"""
+
+    COMMAND = "`changelog`"
+    ALIASES = "`cl`"
+    MESSAGE = partial(HandleInput, "github davep hike ChangeLog.md")
 
 
 ### general.py ends here