+++ src/hike/widgets/command_line/widget.py
@@ -32,6 +32,7 @@ from .base_command import InputCommand
 from .change_directory import ChangeDirectoryCommand
 from .general import (
     BookmarksCommand,
+    ChangeLogCommand,
     ContentsCommand,
     HelpCommand,
     HistoryCommand,
@@ -62,6 +63,7 @@ COMMANDS: Final[tuple[type[InputCommand], ...]] = (
     HelpCommand,
     HistoryCommand,
     LocalCommand,
+    ChangeLogCommand,
     OpenFromBitbucket,
     OpenFromCodeberg,
     OpenFromGitHub,